requirements.txt
torch==2.2.0
torchaudio==2.2.0
librosa==0.10.1
numpy==1.26.4
pandas==2.2.0
scikit-learn==1.4.0
soundfile==0.12.1
audiomentations==0.33.0
matplotlib==3.8.2
seaborn==0.13.2
captum==0.7.0
shap==0.44.0
gradio==4.19.2
pyyaml==6.0.1
tqdm==4.66.2
jupyter==1.0.0
scipy==1.12.0
